create AbstractHydrator
