+++ src/Bridge/ApiPlatform/Hydrator/CommandHydrator.php
@@ -3,31 +3,14 @@
 namespace Botilka\Bridge\ApiPlatform\Hydrator;
 
 use Botilka\Application\Command\Command;
-use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
-use Symfony\Component\Validator\Validator\ValidatorInterface;
 
-final class CommandHydrator implements CommandHydratorInterface
+/**
+ * Provide type hinting & interface implementation.
+ */
+final class CommandHydrator extends AbstractHydrator implements CommandHydratorInterface
 {
-    private $denormalizer;
-    private $validator;
-
-    public function __construct(DenormalizerInterface $denormalizer, ValidatorInterface $validator)
-    {
-        $this->denormalizer = $denormalizer;
-        $this->validator = $validator;
-    }
-
     public function hydrate($data, string $class): Command
     {
-        /** @var Command $command */
-        $command = $this->denormalizer->denormalize($data, $class);
-
-        $violations = $this->validator->validate($command);
-
-        if (0 < \count($violations)) {
-            throw new HydrationException($violations);
-        }
-
-        return $command;
+        return $this->doHydrate($data, $class);
     }
 }
